Типизация order actions
