+++ src/components/order-item/order-item.tsx
@@ -5,7 +5,7 @@ import {Identifier} from 'dnd-core';
 import {ConstructorElement, DragIcon} from "@ya.praktikum/react-developer-burger-ui-components";
 
 import {deleteIngredient} from "../../services/actions/order";
-import {SelectedIngredientType} from "../../services/types";
+import {SelectedIngredientType} from "../../services/types/data";
 import styles from "./order-item.module.css";
 
 interface IOrderItemProps {
