+++ src/services/types/index.ts
@@ -1,9 +1,3 @@
-import {IngredientType} from "./data";
-
-export type SelectedIngredientType = IngredientType & {
-    id: string;
-}
-
 export type LoginType = {
     email: string;
     password: string;
